requirements.txt
pytest==9.0.2
pytest-playwright==0.7.2
playwright==1.40.0
pytest-html==4.2.0
allure-pytest==2.15.3
allure-python-commons==2.15.3
pytest-base-url==2.1.0

